add comments

+++ routes.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request, redirect, jsonify, url_for
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, CategoryItem
+from database_setup import Base, Category, CategoryItem, Users
 from flask import session as login_session
 import random
 import string
@@ -12,8 +12,6 @@ import json
 from flask import make_response, Session, flash
 import requests
 
-
-from flask import Flask
 app = Flask(__name__)
 
 CLIENT_ID = json.loads(
@@ -29,13 +27,15 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-'''
-Appwise Routes
-'''
-
-
+'''Appwise Routes'''
 @app.route('/')
 def index():
+    """index
+    This function helps to show the Category and recent added items,
+    This is the start point to the application
+    Returns:
+        Template : catalog_index.html
+    """
     categories = session.query(Category).order_by(asc(Category.name))
     items = session.query(CategoryItem).order_by(
         (CategoryItem.created_date.desc()))
@@ -43,39 +43,30 @@ def index():
 
 
 '''Auth Routes'''
-
-
 @app.route('/login', methods=['GET', 'POST'])
 def login_form():
+    """login_form
+    This functtion shows the login page if the user is not logged in already.
+    Returns:
+        Template : login.html
+    """
     if request.method == "GET":
         # Create anti-forgery state token
         state = ''.join(random.choice(string.ascii_uppercase +
                                       string.digits) for x in range(32))
         login_session['state'] = state
         return render_template('login.html', STATE=state)
-    elif request.method == "POST":
-        return "Login Handler"
-
-    # # Check that the access token is valid.
-    # access_token = credentials.access_token
-    # url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-    #        % access_token)
-    # h = httplib2.Http()
-    # result = json.loads(h.request(url, 'GET')[1])
-    # # If there was an error in the access token info, abort.
-    # if result.get('error') is not None:
-    #     response = make_response(json.dumps(result.get('error')), 500)
-    #     response.headers['Content-Type'] = 'application/json'
-    #     return response
 
 
-'''
-Catalog Routes
-'''
-
-
+'''Catalog Routes'''
 @app.route('/catalog.json')
 def catalog_json():
+    """catalog_json
+    This functtion return the JSON of all Category information,
+    and all item information in each category.
+    Returns:
+        JSON : [Category][Items]
+    """
     result = {'Category': []}
 
 # Get all Categories
@@ -104,6 +95,11 @@ def catalog_json():
 
 @app.route('/catalog/<path:item_name>/JSON')
 def item_json(item_name):
+    """item_json
+    This functtion return the JSON one item information.
+    Returns:
+        JSON : Item
+    """
 
     item = session.query(CategoryItem).filter_by(name=item_name).one()
     item_obj = {}
@@ -116,22 +112,34 @@ def item_json(item_name):
 
 @app.route('/catalog/<path:cat_name>/items')
 def catalog_index(cat_name):
+    """catalog_index
+    This functtion shows all items in one category.
+    Returns:
+        Template : show.html
+    """
     print("yahoooo")
     category = session.query(Category).filter_by(name=cat_name).one()
     items = session.query(CategoryItem).filter_by(category_id=category.id).all()
     return render_template('category/show.html', category=category, items=items)
 
 
-'''
-Item Routes
-'''
-
-
+'''Item Routes'''
 @app.route('/catalog/<path:item_name>/edit', methods=['GET', 'POST'])
 def item_edit(item_name):
+    """Item Edit
+    This Function helps to show the edit form and update the item with new values.
+    Returns:
+        Redirect|Template -- 
+        If get request > Template form.html
+        If Post request > redirect url_for('index')
+    """
     if 'username' not in login_session:
         return redirect('/login')
     item = session.query(CategoryItem).filter_by(name=item_name).one()
+
+    if login_session['id'] != item.user_id:
+        # Not Authorised
+        return redirect('/login')
 
     if request.method == 'GET':
         categories = session.query(Category).order_by(asc(Category.name))
@@ -152,22 +160,30 @@ def item_edit(item_name):
 
 @app.route('/catalog/create', methods=['GET', 'POST'])
 def item_create():
+    """item_create
+    This Function helps to show the create form and create the item in the database.
+    Returns:
+        Redirect|Template -- 
+        If get request > Template create.html
+        If Post request > redirect url_for('index')
+    """
     if 'username' not in login_session:
+        # Not Authenticated
         return redirect('/login')
+
 
     if request.method == 'GET':
         categories = session.query(Category).order_by(asc(Category.name))
         return render_template('items/create.html', categories=categories)
     elif request.method == "POST":
-        print("yahoooooooo")
         if request.form['name'] and request.form['description'] and request.form['category_id']:
             print("in if")
-            newITem = CategoryItem(user_id=1, user_email=login_session['email'], name=request.form['name'],
+            currentUser = session.query(Users).filter_by(email= login_session['email']).one()
+            newITem = CategoryItem(user_id=currentUser.id, user_email=login_session['email'], name=request.form['name'],
                                    description=request.form['description'], category_id=request.form['category_id'])
             session.add(newITem)
             session.commit()
             flash("Item added")
-
         else:
             flash("Item not added!")
 
@@ -176,11 +192,22 @@ def item_create():
 
 @app.route('/catalog/<path:item_name>/delete', methods=['GET', 'POST'])
 def item_delete(item_name):
+    """item_delete
+    This Function helps to show the Delete form and delelte the item from the database.
+    Returns:
+        Redirect|Template -- 
+        If get request > Template delete.html
+        If Post request > redirect url_for('index')
+    """
 
     if 'username' not in login_session:
         return redirect('/login')
-
     item = session.query(CategoryItem).filter_by(name=item_name).one()
+
+    if login_session['id'] != item.user_id:
+        # Not Authorised
+        return redirect('/login')
+    
     if request.method == 'POST':
         print("wrong")
         session.delete(item)
@@ -192,12 +219,23 @@ def item_delete(item_name):
 
 @app.route('/catalog/<path:item_name>/show')
 def show(item_name):
+    """show
+    This Function helps to show the item detail form.
+    Returns:
+        Template -- show.html
+    """
     item = session.query(CategoryItem).filter_by(name=item_name).one()
     return render_template('items/show.html', item=item)
 
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
+    """gconnect
+    This Function helps to login with google.
+    Returns:
+        If user logged in already,
+        OR user logged in successfully > redirect to index.
+    """
     # Validate state token
     print("<gco>")
     if request.args.get('state') != login_session['state']:
@@ -265,13 +303,30 @@ def gconnect():
 
     data = answer.json()
 
-    print(data)
-
-    login_session['username'] = data['email']
+    login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
     login_session['uid'] = data['id']
 
+    existUser = session.query(Users).filter_by(email=login_session['email']).first()
+
+    if existUser:
+        print("Exist user, so update info")
+        existUser.name = login_session['username']
+        existUser.email = login_session['email']
+        existUser.picture = login_session['email']
+        session.add(existUser)
+        session.commit()
+    else:
+        print("create new user")
+        newUser = Users(name=login_session['username'], email=login_session['email'],
+                picture= login_session['email'])
+        session.add(newUser)
+        session.commit()
+
+    existUser = session.query(Users).filter_by(email=login_session['email']).first()
+    login_session['id'] = existUser.id
+      
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -279,7 +334,7 @@ def gconnect():
     output += '<img src="'
     output += login_session['picture']
     output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-    # flash("you are now logged in as %s" % login_session['username'])
+    flash("you are now logged in as %s" % login_session['username'])
     print("done!")
     return output
 
@@ -288,6 +343,11 @@ def gconnect():
 
 @app.route('/logout')
 def logout():
+    """logout
+    This Function helps to logout from google.
+    Returns:
+        redirect to index
+    """
     access_token = login_session['access_token']
     username = login_session['username']
     print('In gdisconnect access token is %s', access_token)
@@ -319,10 +379,5 @@ def logout():
 if True or __name__ == '__main__':
     app.secret_key = 'super_secret_key'
     app.debug = True
-
     app.config['SESSION_TYPE'] = 'filesystem'
-
-    # sess = Session()
-    # sess.init_app(app)
-
     app.run(host='0.0.0.0', port=8000)
